refactor(types): pagination
